+++ core/src/main/scala/org/apache/spark/sql/crossdata/XDFunctionRegistry.scala
@@ -20,8 +20,7 @@ import org.apache.spark.Logging
 import org.apache.spark.sql.AnalysisException
 import org.apache.spark.sql.catalyst.analysis.FunctionRegistry
 import org.apache.spark.sql.catalyst.analysis.FunctionRegistry.FunctionBuilder
-import org.apache.spark.sql.catalyst.expressions.Expression
-import org.apache.spark.sql.catalyst.expressions.ExpressionInfo
+import org.apache.spark.sql.catalyst.expressions.{Expression, ExpressionInfo}
 
 import scala.util.Try