+++ core/src/main/scala/org/apache/spark/sql/crossdata/execution/datasources/ddl.scala
@@ -17,27 +17,22 @@ package org.apache.spark.sql.crossdata.execution.datasources
 
 import com.stratio.crossdata.connector.TableInventory
 import org.apache.spark.Logging
+
 import org.apache.spark.sql.catalyst.TableIdentifier
 import org.apache.spark.sql.catalyst.expressions.Attribute
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
-import org.apache.spark.sql.crossdata.CrossdataTable
-import org.apache.spark.sql.crossdata.XDCatalog
+import org.apache.spark.sql.crossdata.{CrossdataTable, XDCatalog}
 import org.apache.spark.sql.execution.RunnableCommand
-import org.apache.spark.sql.execution.datasources.LogicalRelation
-import org.apache.spark.sql.execution.datasources.ResolvedDataSource
+import org.apache.spark.sql.execution.datasources.{ResolvedDataSource, LogicalRelation}
 import org.apache.spark.sql.sources.RelationProvider
-import org.apache.spark.sql.types.ArrayType
-import org.apache.spark.sql.types.BooleanType
-import org.apache.spark.sql.types.StringType
-import org.apache.spark.sql.types.StructField
-import org.apache.spark.sql.types.StructType
-import org.apache.spark.sql.AnalysisException
-import org.apache.spark.sql.Row
-import org.apache.spark.sql.SQLContext
+
+import org.apache.spark.sql.types.{ArrayType, BooleanType, StringType, StructField, StructType}
+import org.apache.spark.sql.{AnalysisException, Row, SQLContext}
+
 import XDCatalog._
 
 
-private[crossdata] case class ImportTablesUsingWithOptions(datasource: String, opts: Map[String, String])
+private [crossdata] case class ImportTablesUsingWithOptions(datasource: String, opts: Map[String, String])
   extends LogicalPlan with RunnableCommand with Logging {
 
   // The result of IMPORT TABLE has only tableIdentifier so far.