+++ driver/src/main/scala/com/stratio/crossdata/driver/querybuilder/dslentities/combination.scala
@@ -15,7 +15,8 @@
  */
 package com.stratio.crossdata.driver.querybuilder.dslentities
 
-import com.stratio.crossdata.driver.querybuilder.{CrossdataSQLStatement, RunnableQuery}
+import com.stratio.crossdata.driver.querybuilder.CrossdataSQLStatement
+import com.stratio.crossdata.driver.querybuilder.RunnableQuery
 
 
 object CombineType extends Enumeration {
@@ -27,7 +28,7 @@ object CombineType extends Enumeration {
 }
 
 
-import CombineType._
+import com.stratio.crossdata.driver.querybuilder.dslentities.CombineType.CombineType
 
 case class CombinationInfo(combineType: CombineType, runnableQuery: RunnableQuery) extends CrossdataSQLStatement {
   override private[querybuilder] def toXDQL: String = s" ${combineType.toString} ${runnableQuery.toXDQL}"