+++ CHANGELOG.md
@@ -2,8 +2,6 @@
 
 Only listing significant user-visible, not internal code cleanups and minor bug fixes. 
 
-## 1.1.0 (Upcoming)
-
 ## 1.0.0
 
 * Architecture migrated to Spark Catalyst.
@@ -14,5 +12,6 @@ Only listing significant user-visible, not internal code cleanups and minor bug 
 * Improved and extended SparkContext: XDContext.
 * Pluggable and persistent catalog.
 * Metadata discovery.
+* Temporary views.
 * P2P architecture: High availability and scalability.
 * Query builder.